feat: handle stray CLAUDE.md in ion migrate without target gate
Migrating a project that has CLAUDE.md but no skills history previously
required users to either configure claude as a target first or live with
the misleading "No skills found" output. Drop the target gate — the
existence of a regular CLAUDE.md is signal enough to convert it to a
symlink to AGENTS.md.

When the AGENTS.md step did real work but no skills were discovered,
suppress the "scanning directories" / "No skills found" chatter and emit
a focused "Migration complete!" with a git commit. The empty-state
message becomes "Nothing to migrate." so the negative wording only
appears when there really is nothing to do.

Extracts the CLAUDE.md migration helper into commands/agents.rs so it has
a coherent home alongside other AGENTS.md logic; commands/migrate.rs now
delegates to it.
